requirements.txt
fastapi>=0.127.0
uvicorn>=0.34.0
pydantic>=2.12.0
scikit-learn>=1.8.0
joblib>=1.5.0
numpy>=2.0.0
pandas>=2.3.0
motor>=3.3.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
email-validator